Move Abspath function to headscale utils
